# [perf] do not propagate constant tag addition