It seems we use `Tracker.add` only inside `consumeTag`, and it already has null check, I think it's fine to move `CONSTANT_TAG` check to this level to avoid extra function call chain