@lifeart 2023-12-23
```
   Benchmark Results Summary    

duration phase no difference [-25ms to 34ms]
render1000Items1End phase no difference [-1ms to 8ms]
clearItems1End phase no difference [0ms to 1ms]
render1000Items2End phase no difference [-4ms to 1ms]
clearItems2End phase no difference [0ms to 0ms]
render10000Items1End phase no difference [-17ms to 19ms]
clearItems3End phase estimated regression +2ms [1ms to 3ms] OR +1.15% [0.58% to 1.61%]
render1000Items3End phase no difference [-5ms to 5ms]
append1000Items1End phase no difference [-6ms to 1ms]
updateEvery10thItem1End phase no difference [0ms to 0ms]
selectFirstRow1End phase no difference [-1ms to 0ms]
selectSecondRow1End phase no difference [0ms to 1ms]
removeFirstRow1End phase no difference [-1ms to 2ms]
removeSecondRow1End phase no difference [-2ms to 0ms]
swapRows1End phase no difference [0ms to 1ms]
clearItems4End phase no difference [0ms to 1ms]
paint phase no difference [0ms to 0ms]
```